extract DrtQSimModule from DrtControlerCreator
