+++ contribs/drt/src/main/java/org/matsim/contrib/drt/run/DrtControlerCreator.java
@@ -29,45 +29,20 @@ import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.api.core.v01.TransportMode;
 import org.matsim.contrib.drt.analysis.DrtAnalysisModule;
-import org.matsim.contrib.drt.optimizer.DefaultDrtOptimizer;
 import org.matsim.contrib.drt.optimizer.DrtOptimizer;
-import org.matsim.contrib.drt.optimizer.VehicleData;
-import org.matsim.contrib.drt.optimizer.VehicleDataEntryFactoryImpl;
 import org.matsim.contrib.drt.optimizer.insertion.DefaultUnplannedRequestInserter;
 import org.matsim.contrib.drt.optimizer.insertion.ParallelPathDataProvider;
-import org.matsim.contrib.drt.optimizer.insertion.PrecalculablePathDataProvider;
-import org.matsim.contrib.drt.optimizer.insertion.UnplannedRequestInserter;
-import org.matsim.contrib.drt.passenger.DrtRequestCreator;
 import org.matsim.contrib.drt.routing.DrtRoute;
 import org.matsim.contrib.drt.routing.DrtRouteFactory;
 import org.matsim.contrib.drt.routing.DrtStageActivityType;
-import org.matsim.contrib.drt.schedule.DrtTaskFactory;
-import org.matsim.contrib.drt.schedule.DrtTaskFactoryImpl;
-import org.matsim.contrib.drt.scheduler.DrtScheduleInquiry;
-import org.matsim.contrib.drt.scheduler.DrtScheduleTimingUpdater;
-import org.matsim.contrib.drt.scheduler.EmptyVehicleRelocator;
-import org.matsim.contrib.drt.scheduler.RequestInsertionScheduler;
-import org.matsim.contrib.drt.vrpagent.DrtActionCreator;
-import org.matsim.contrib.dvrp.optimizer.VrpOptimizer;
-import org.matsim.contrib.dvrp.passenger.PassengerEngine;
-import org.matsim.contrib.dvrp.passenger.PassengerRequestCreator;
 import org.matsim.contrib.dvrp.run.DvrpModule;
-import org.matsim.contrib.dvrp.run.MobsimTimerProvider;
-import org.matsim.contrib.dvrp.trafficmonitoring.DvrpTravelDisutilityProvider;
-import org.matsim.contrib.dvrp.vrpagent.VrpAgentLogic.DynActionCreator;
 import org.matsim.contrib.otfvis.OTFVisLiveModule;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.groups.PlanCalcScoreConfigGroup.ActivityParams;
 import org.matsim.core.config.groups.PlanCalcScoreConfigGroup.ModeParams;
 import org.matsim.core.controler.Controler;
-import org.matsim.core.mobsim.framework.MobsimTimer;
-import org.matsim.core.mobsim.qsim.AbstractQSimModule;
 import org.matsim.core.population.routes.RouteFactories;
 import org.matsim.core.scenario.ScenarioUtils;
-
-import com.google.inject.Key;
-import com.google.inject.name.Named;
-import com.google.inject.name.Names;
 
 /**
  * @author jbischoff
@@ -124,7 +99,7 @@ public final class DrtControlerCreator {
 
 	public static void addDrtToControler(Controler controler) {
 		String mode = DrtConfigGroup.get(controler.getConfig()).getMode();
-		controler.addQSimModule(DrtControlerCreator.createQSimModule());
+		controler.addQSimModule(new DrtQSimModule());
 		controler.addOverridingModule(DvrpModule.createModule(mode,
 				Arrays.asList(DrtOptimizer.class, DefaultUnplannedRequestInserter.class,
 						ParallelPathDataProvider.class)));
@@ -165,40 +140,5 @@ public final class DrtControlerCreator {
 
 		config.addConfigConsistencyChecker(new DrtConfigConsistencyChecker());
 		config.checkConsistency();
-	}
-
-	public static AbstractQSimModule createQSimModule() {
-		return new AbstractQSimModule() {
-			@Override
-			protected void configureQSim() {
-				bind(MobsimTimer.class).toProvider(MobsimTimerProvider.class).asEagerSingleton();
-				DvrpTravelDisutilityProvider.bindTravelDisutilityForOptimizer(binder(),
-						DefaultDrtOptimizer.DRT_OPTIMIZER);
-
-				bind(DrtOptimizer.class).to(DefaultDrtOptimizer.class).asEagerSingleton();
-
-				bind(DefaultUnplannedRequestInserter.class).asEagerSingleton();
-				bind(UnplannedRequestInserter.class).to(DefaultUnplannedRequestInserter.class);
-				bind(VehicleData.EntryFactory.class).to(VehicleDataEntryFactoryImpl.class).asEagerSingleton();
-
-				bind(DrtTaskFactory.class).to(DrtTaskFactoryImpl.class).asEagerSingleton();
-
-				bind(EmptyVehicleRelocator.class).asEagerSingleton();
-				bind(DrtScheduleInquiry.class).asEagerSingleton();
-				bind(RequestInsertionScheduler.class).asEagerSingleton();
-				bind(DrtScheduleTimingUpdater.class).asEagerSingleton();
-
-				bind(ParallelPathDataProvider.class).asEagerSingleton();
-				bind(PrecalculablePathDataProvider.class).to(ParallelPathDataProvider.class);
-
-				Named modeNamed = Names.named(DrtConfigGroup.get(getConfig()).getMode());
-				bind(VrpOptimizer.class).annotatedWith(modeNamed).to(DrtOptimizer.class);
-				bind(DynActionCreator.class).annotatedWith(modeNamed).to(DrtActionCreator.class).asEagerSingleton();
-				bind(PassengerRequestCreator.class).annotatedWith(modeNamed)
-						.to(DrtRequestCreator.class)
-						.asEagerSingleton();
-				bind(PassengerEngine.class).annotatedWith(Drt.class).to(Key.get(PassengerEngine.class, modeNamed));
-			}
-		};
 	}
 }
